prevented compiler to show some warnings

+++ include/axolote/entity.hpp
@@ -25,7 +25,7 @@ namespace axolote
 
     public:
         Entity();
-        void set_matrix(int idx, glm::mat4 mat = glm::mat4(1.0f));
+        void set_matrix(size_t idx, glm::mat4 mat = glm::mat4(1.0f));
         void draw(Shader shader);
         void add_model(Model m, glm::mat4 mat = glm::mat4(1.0f));
         void add_mesh(Mesh m, glm::mat4 mat = glm::mat4(1.0f));
